Fix of issue 11. Compiles on x86_64 again.
git-svn-id: http://netsniff-ng.googlecode.com/svn/trunk@379 21e0ff64-9a0b-11de-825e-994487f65616

+++ netsniff-ng/include/netsniff-ng/ether_types.h
@@ -23,7 +23,7 @@
 #include <stdint.h>
 
 typedef struct {
-	uint32_t id;
+	uintptr_t id;
 	char *type;
 } ether_type_t;
