+++ netsniff-ng/include/netsniff-ng/ports_udp.h
@@ -23,7 +23,7 @@
 #include <stdint.h>
 
 typedef struct {
-	uint32_t id;
+	uintptr_t id;
 	char *port;
 } port_udp_t;
